Add files via upload
The program is a crawler which searches Microsoft Bing for a certain keyword and plots a choropleth map of Italy's provinces, after comparing the result with a list of local newspapers to establish the origin of the instance. If the news is from a newspaper from Treviso, say, than the program will count one for that province. 
Additionally, the program will search for any numbers in the news feed and plot a frequency histograms of the found instances.